[feat/#6] 필수 과제 추가
